refactor out some shared config stuffu
